Save difference between db & Riot in admin area
In the admin area, if there is a difference between what is in the
database for a particular model and what is from Riot, show the option
to persist the difference. Will use the persist API through AJAX POST
requests. Also include the option to truncate the table. Use server side
processing for DataTables.