Modify query of Items for Troll Build
Previously, the Map<String, Boolean> of all game maps for a Item could
be null. In addition, the map ID of 1 (SummonersRift) shouldn't be true
but now it is never false in any items returned from the API. The new
map ID to look for is 11 (SummonersRiftNew), and this one should not be
false for troll build items.