Proxy API calls to REST backend
In dev, with ng serve and spring boot run, CORS is an issue with cross domain requests. In prod, spring boot will serve a built Angular bundle and both will be on the same domain at that time. This also eliminates a need for a base url prepending http interceptor. So battling CORS is no longer needed although I had just figured it all out...