Remove firefox and ie karma launchers
Maybe temporarily removed. But it seems with the new es5BrowserSupport feature and running tests with karma, es2015-polyfills script is not added to the karma-context.html like it would be added to the index.html. Therefore IE karma tests fail to run due to missing polyfills.

In the future, Firefox and IE will be used in e2e tests instead of unit tests.